Move Pixiu com files in /cmd/pixiu to pkg/cmd
